+++ common/include/pcl/common/impl/centroid.hpp
@@ -688,7 +688,7 @@ demeanPointCloud (const pcl::PointCloud<PointT> &cloud_in,
   }
   else
   {
-    cloud_out.width    = static_cast<std::uint32_t> (indices.size ());
+    cloud_out.width    = indices.size ();
     cloud_out.height   = 1;
   }
   cloud_out.resize (indices.size ());
