+++ examples/segmentation/example_extract_clusters_normals.cpp
@@ -89,7 +89,7 @@ main (int, char **argv)
     pcl::PointCloud<pcl::PointXYZ>::Ptr cloud_cluster (new pcl::PointCloud<pcl::PointXYZ>);
     for (const int &index : it->indices)
       cloud_cluster->push_back ((*cloud_ptr)[index]); 
-    cloud_cluster->width = static_cast<std::uint32_t> (cloud_cluster->size ());
+    cloud_cluster->width = cloud_cluster->size ();
     cloud_cluster->height = 1;
     cloud_cluster->is_dense = true;
