+++ keypoints/include/pcl/keypoints/impl/iss_3d.hpp
@@ -447,7 +447,7 @@ pcl::ISSKeypoint3D<PointInT, PointOutT, NormalT>::detectKeypoints (PointCloudOut
   }
 
   output.header = input_->header;
-  output.width = static_cast<std::uint32_t> (output.size ());
+  output.width = output.size ();
   output.height = 1;
 
   // Clear the contents of variables and arrays before the beginning of the next computation.
