+++ tools/voxel_grid_occlusion_estimation.cpp
@@ -205,7 +205,7 @@ int main (int argc, char** argv)
   print_info ("[done, "); print_value ("%g", tt.toc ()); print_info (" ms : "); print_value ("%d", (int)occluded_voxels.size ()); print_info (" occluded voxels]\n");
   
   CloudT::Ptr occ_centroids (new CloudT);
-  occ_centroids->width = static_cast<int> (occluded_voxels.size ());
+  occ_centroids->width = occluded_voxels.size ();
   occ_centroids->height = 1;
   occ_centroids->is_dense = false;
   occ_centroids->points.resize (occluded_voxels.size ());
@@ -220,7 +220,7 @@ int main (int argc, char** argv)
   }
 
   CloudT::Ptr cloud_centroids (new CloudT);
-  cloud_centroids->width = static_cast<int> (input_cloud->size ());
+  cloud_centroids->width = input_cloud->size ();
   cloud_centroids->height = 1;
   cloud_centroids->is_dense = false;
   cloud_centroids->points.resize (input_cloud->size ());
